:checkered_flag: Fixed constants
